requirements.txt
google-cloud-aiplatform==1.7.0
kfp==1.8.9
google-cloud-pipeline-components==0.2.0
pandas
numpy